Move Content-Disposition helpers to gluon.http and make them public
