+++ gluon/sqlhtml.py
@@ -32,7 +32,7 @@ from pydal.helpers.methods import _repr_ref, bar_encode, merge_tablemaps, smart_
 from pydal.objects import Expression, Field, Row, Rows, Set, Table
 
 import gluon.serializers as serializers
-from gluon.globals import _content_disposition_header, current
+from gluon.globals import current
 from gluon.html import (
     BR,
     CAT,
@@ -67,7 +67,7 @@ from gluon.html import (
     XmlComponent,
     truncate_string,
 )
-from gluon.http import HTTP, redirect
+from gluon.http import HTTP, content_disposition_header, redirect
 from gluon.storage import Storage
 from gluon.utils import md5_hash
 from gluon.validators import (
@@ -3164,7 +3164,7 @@ class SQLFORM(FORM):
                 # http layer only strips CR/LF, so `"` and `;` would survive.
                 # Reuse the shared Content-Disposition filename helper.
                 response.headers["Content-Type"] = oExp.content_type
-                response.headers["Content-Disposition"] = _content_disposition_header(
+                response.headers["Content-Disposition"] = content_disposition_header(
                     "%s.%s" % (export_filename, oExp.file_ext)
                 )
                 raise HTTP(200, oExp.export(), **response.headers)
